+++ packages/js/src/BaseRoomSession.ts
@@ -10,6 +10,7 @@ import {
   validateEventsToSubscribe,
   EventEmitter,
   SDKWorker,
+  selectors,
 } from '@signalwire/core'
 import {
   getDisplayMedia,
@@ -93,6 +94,14 @@ export class RoomSessionConnection
     if (options.eventsWatcher) {
       this.runWorker('eventsWatcher', { worker: options.eventsWatcher })
     }
+  }
+
+  get self() {
+    return this.select(selectors.getSelf)
+  }
+
+  get target() {
+    return this.select(selectors.getTarget)
   }
 
   get screenShareList() {
