+++ packages/js/src/utils/interfaces.ts
@@ -37,6 +37,7 @@ import type {
   VideoRoomDeviceDisconnectedEventNames,
   DeviceDisconnectedEventParams,
   VideoRoomDeviceEventNames,
+  InternalUnifiedActionTarget,
 } from '@signalwire/core'
 import { INTERNAL_MEMBER_UPDATABLE_PROPS } from '@signalwire/core'
 import type { MediaEvent } from '@signalwire/webrtc'
@@ -197,6 +198,8 @@ export interface MemberCommandWithValueParams extends MemberCommandParams {
 export interface BaseRoomInterface {
   join(): Promise<unknown>
   leave(): Promise<unknown>
+  self?: InternalUnifiedActionTarget,
+  target?: InternalUnifiedActionTarget
 }
 
 export interface LocalOverlay {
